Litecoin: Disable PoW Sanity check while loading block index from disk.
We use the sha256 hash for the block index for performance reasons, which is recorded for later use.
CheckProofOfWork() uses the scrypt hash which is discarded after a block is accepted.
While it is technically feasible to verify the PoW, doing so takes several minutes as it
requires recomputing every PoW hash during every Litecoin startup.
We opt instead to simply trust the data that is on your local disk.